based on commit from Mon, 30 Nov 2009 18:25:26 +0000 by danieljames <danieljames@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Use consistent names for template parameters.

I'm trying to fix the codegear ICEs, but it's hard to tell the cause.
Since the error happens operator== I suspect it's either to do
with defining friend functions with different template names or
something to do with friend functions in general. This is the first stab
in the dark at fixing this.

git-svn-id: http://svn.boost.org/svn/boost/trunk@58062 b8fc166d-592f-0410-95f2-cb63ce0dd405