based on commit from Fri, 4 Dec 2009 00:51:50 +0000 by danieljames <danieljames@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Workaround codegear ICE.

It seems that the problem is calling sizeof on a dependent type when the
containers have only been used by reference. So by putting in these
dummy structures with the containers as members, it helps the compiler
instantiate the class to the level where sizeof works. I hope.

git-svn-id: http://svn.boost.org/svn/boost/trunk@58130 b8fc166d-592f-0410-95f2-cb63ce0dd405