based on commit from Thu, 7 Jan 2010 05:57:33 +0000 by noel_belcourt <noel_belcourt@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
This patch ensures we close both file descriptors following the dup2 calls. This is mostly for completeness but also because I've occasionally noticed output isn't flushed from a terminated process (when it is killed for exceeding the time limit).  I'm sure this won't break anything and am slightly hopeful that it might ensure output is flushed before the terminating process is killed.

git-svn-id: http://svn.boost.org/svn/boost/trunk@58785 b8fc166d-592f-0410-95f2-cb63ce0dd405