based on commit from Fri, 5 Feb 2010 11:51:46 +0000 by vladimir_prus <vladimir_prus@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Use the pathf90 compiler for all compilation.

Patch from Spencer E. Olson.

Here's the explanation from email:

The version of pathscale that I have access to is 3.2 and there doesn't seem
to be a pathf77 executable.  The man pages do not reference it, and there is
also not a -f77 option that specifies that the code is fortran 77.  Rather,
the man pages indicate that pathscale only recognizes the difference between
free form and fixed form fortran instead of F77 vs F90 vs F95.

Does anyone have an available installation that differs from this?  If not,
I'll submit another small patch for changing this to using pathf90 to compile
f77 code--I've tested it and it appears to work well enough.

git-svn-id: http://svn.boost.org/svn/boost/trunk@59508 b8fc166d-592f-0410-95f2-cb63ce0dd405