based on commit from Mon, 8 Feb 2010 15:18:57 +0000 by vladimir_prus <vladimir_prus@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Make sure we use the same name of project module everywhere.

In particular, this fixes problems on windows where project.load
may be passed a path from Jamfile, and then load-jamfile may use
a path using the exact case as stored in the file system -- different
from the path passed to project.load.

git-svn-id: http://svn.boost.org/svn/boost/trunk@59581 b8fc166d-592f-0410-95f2-cb63ce0dd405