based on commit from Mon, 22 Mar 2010 15:26:45 +0000 by grafik <grafik@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Fix possible problems when trailing or leading whitespace is added to the env vars used, for example CC and CFLAGS when using the cc toolset. Even though I can't get compilation to fail even with the extra whitespace during building it does seem to fail for some compilers that don't ignore the whitespace arguments. (Fixes #3630)

git-svn-id: http://svn.boost.org/svn/boost/trunk@60768 b8fc166d-592f-0410-95f2-cb63ce0dd405