based on commit from Sun, 4 Apr 2010 18:45:45 +0000 by andreas_huber69 <andreas_huber69@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Fixes #3963. post_event() is now protected and forwards to post_event_impl(), which unfortunately must remain public. So, client code can still abuse the class by calling post_event_impl(), but the abuse should now be more obvious. I considered adding an assert to post_event_impl(), so that unauthorized calls from the outside could at least be detected at runtime, but it turned out that an assert requires more additional machinery than expected, which is why it was not implemented.

Thanks for the report!

git-svn-id: http://svn.boost.org/svn/boost/trunk@61048 b8fc166d-592f-0410-95f2-cb63ce0dd405