based on commit from Wed, 21 Apr 2010 17:19:06 +0000 by hkaiser <hkaiser@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
MPL: fixed #4061: gcc-4.5 compilation problems related to arity_helper, applied attached patch to the main aux_/template_arity.hpp and regenerated the corresponding file preprocessed/gcc/template_arity.hpp. No other preprocessed files are affected.

This patch seemed to be fine as all it does is to qualify an internal name in order to avoid it being looked up through ADL. This appears to be reasonably safe as this name is internal and not supposed to be found using ADL in the first place.

git-svn-id: http://svn.boost.org/svn/boost/trunk@61467 b8fc166d-592f-0410-95f2-cb63ce0dd405