based on commit from Sat, 1 May 2010 00:13:03 +0000 by dgregor <dgregor@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
At several places in the ptr_container tests, we have the following pattern:

void f(const std::auto_ptr<int> &); void g() {

f(std::auto_ptr<int>(new int(5)));

}

This is code is ill-formed in C++98/03, because the compiler is
required to check there there exits a copy constructor that would be
able to make a copy of the auto_ptr before binding it to a reference,
even if that copy won't actually be made. Not all compilers diagnose
this error, but Clang and EDG in strict mode do diagnose the problem.

Fixes #4174.

git-svn-id: http://svn.boost.org/svn/boost/trunk@61718 b8fc166d-592f-0410-95f2-cb63ce0dd405