based on commit from Sun, 27 Jun 2010 16:39:36 +0000 by danieljames <danieljames@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Don't complain about an id immediately followed by a name.

Some documentation uses markup of the form:

{{{
<div id="foo"><h1><a name="foo"></a>Foo</h1>
...
</div>
}}}

This was appearing as a duplicate bookmark, but it's valid. So allow an
`id` immediately followed by an identical `name`. Could allow some
problematic markup but it's the best I could come up with without
a html parser.

git-svn-id: http://svn.boost.org/svn/boost/trunk@63389 b8fc166d-592f-0410-95f2-cb63ce0dd405