based on commit from Mon, 19 Jul 2010 13:20:51 +0000 by vladimir_prus <vladimir_prus@b8fc166d-592f-0410-95f2-cb63ce0dd405>:
Implemented a mechanism for parameter translation from bjam to python signatures.

This allows to expose the same interface to Jamfiles as before, e.g.

     rule toolset.flags ( rule variable condition * : values * )

and use natural interface in the Python side, where, 'rule', 'variable',
'condition' and 'values' are parameters, and the first two of them
will get string, not list of strings.

git-svn-id: http://svn.boost.org/svn/boost/trunk@64158 b8fc166d-592f-0410-95f2-cb63ce0dd405